Added an integration test
