+++ src/commands/stat.rs
@@ -1,4 +1,4 @@
-use crate::lib::system::{storage, string, tools, util};
+use crate::system::{storage, string, tools, util};
 
 pub fn show() {
   let stats = storage::show_db().expect(string::ERR_SHOW_STATISTICS);
